+++ src/app/blind-creator/blind-creator.component.css
