+++ src/app/blind-creator/blind-creator.component.html
